+++ clinet/src/pages/dashboard/TourSpot/OthersInputField.jsx
@@ -1,0 +1,87 @@
+import { TextField } from "@mui/material";
+
+const OthersInputField = ({ handleChange, formData }) => {
+  return (
+    <>
+      <TextField
+        type="text"
+        multiline
+        onChange={handleChange}
+        value={formData?.perfectTourTime}
+        name="perfectTourTime"
+        className="w-full"
+        id="outlined-basic4"
+        label="perfectTourTime"
+        variant="outlined"
+      />
+      <TextField
+        type="text"
+        multiline
+        onChange={handleChange}
+        value={formData?.howToGoThere}
+        name="howToGoThere"
+        className="w-full"
+        id="outlined-basic4"
+        label="howToGoThere"
+        variant="outlined"
+      />
+      <TextField
+        type="text"
+        multiline
+        onChange={handleChange}
+        value={formData?.howToStayThere}
+        name="howToStayThere"
+        className="w-full"
+        id="outlined-basic4"
+        label="howToStayThere"
+        variant="outlined"
+      />
+      <TextField
+        type="text"
+        multiline
+        onChange={handleChange}
+        value={formData?.howDoHere}
+        name="howDoHere"
+        className="w-full"
+        id="outlined-basic4"
+        label="howDoHere"
+        variant="outlined"
+      />
+      <TextField
+        type="text"
+        multiline
+        onChange={handleChange}
+        value={formData?.whereToEat}
+        name="whereToEat"
+        className="w-full"
+        id="outlined-basic4"
+        label="whereToEat"
+        variant="outlined"
+      />
+      <TextField
+        type="text"
+        multiline
+        onChange={handleChange}
+        value={formData?.tourTipsGuide}
+        name="tourTipsGuide"
+        className="w-full"
+        id="outlined-basic4"
+        label="tourTipsGuide"
+        variant="outlined"
+      />
+      <TextField
+        type="text"
+        multiline
+        onChange={handleChange}
+        value={formData?.topTourPlace}
+        name="topTourPlace"
+        className="w-full"
+        id="outlined-basic4"
+        label="topTourPlace"
+        variant="outlined"
+      />
+    </>
+  );
+};
+
+export default OthersInputField;
