+++ server/models/TourSpot.js
@@ -3,31 +3,26 @@ const mongoose = require("mongoose");
 const TourSpotSchema = new mongoose.Schema({
   name: {
     type: String,
-    required: true,
   },
 
   description: {
     type: String,
-    required: true,
   },
 
   photo: {
     type: String,
-    required: true,
   },
 
   countryId: {
     type: mongoose.Schema.Types.ObjectId,
     ref: "Country",
-    required: true,
   },
   divisionId: {
     type: mongoose.Schema.Types.ObjectId,
     ref: "Division",
-    required: true,
   },
 
-  cityId: { type: mongoose.Schema.Types.ObjectId, ref: "City", required: true },
+  cityId: { type: mongoose.Schema.Types.ObjectId, ref: "City" },
 
   perfectTourTime: { type: String },
